can add blocks, face painting optimized

+++ src/SDL2/i686-w64-mingw32/include/SDL2/SDL_rect.h
@@ -28,10 +28,10 @@
 #ifndef SDL_rect_h_
 #define SDL_rect_h_
 
-#include "SDL_stdinc.h"
 #include "SDL_error.h"
 #include "SDL_pixels.h"
 #include "SDL_rwops.h"
+#include "SDL_stdinc.h"
 
 #include "begin_code.h"
 /* Set up for C function definitions, even when using C++ */
@@ -45,24 +45,21 @@ extern "C" {
  * \sa SDL_EnclosePoints
  * \sa SDL_PointInRect
  */
-typedef struct SDL_Point
-{
-    int x;
-    int y;
+typedef struct SDL_Point {
+  int x;
+  int y;
 } SDL_Point;
 
 /**
- * The structure that defines a point (floating point)
+ * The structure that defines a point (doubleing point)
  *
  * \sa SDL_EncloseFPoints
  * \sa SDL_PointInFRect
  */
-typedef struct SDL_FPoint
-{
-    float x;
-    float y;
+typedef struct SDL_FPoint {
+  double x;
+  double y;
 } SDL_FPoint;
-
 
 /**
  * A rectangle, with the origin at the upper left (integer).
@@ -75,15 +72,13 @@ typedef struct SDL_FPoint
  * \sa SDL_UnionRect
  * \sa SDL_EnclosePoints
  */
-typedef struct SDL_Rect
-{
-    int x, y;
-    int w, h;
+typedef struct SDL_Rect {
+  int x, y;
+  int w, h;
 } SDL_Rect;
 
-
 /**
- * A rectangle, with the origin at the upper left (floating point).
+ * A rectangle, with the origin at the upper left (doubleing point).
  *
  * \sa SDL_FRectEmpty
  * \sa SDL_FRectEquals
@@ -95,39 +90,39 @@ typedef struct SDL_Rect
  * \sa SDL_EncloseFPoints
  * \sa SDL_PointInFRect
  */
-typedef struct SDL_FRect
-{
-    float x;
-    float y;
-    float w;
-    float h;
+typedef struct SDL_FRect {
+  double x;
+  double y;
+  double w;
+  double h;
 } SDL_FRect;
-
 
 /**
  * Returns true if point resides inside a rectangle.
  */
-SDL_FORCE_INLINE SDL_bool SDL_PointInRect(const SDL_Point *p, const SDL_Rect *r)
-{
-    return ( (p->x >= r->x) && (p->x < (r->x + r->w)) &&
-             (p->y >= r->y) && (p->y < (r->y + r->h)) ) ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_PointInRect(const SDL_Point *p,
+                                          const SDL_Rect *r) {
+  return ((p->x >= r->x) && (p->x < (r->x + r->w)) && (p->y >= r->y) &&
+          (p->y < (r->y + r->h)))
+             ? SDL_TRUE
+             : SDL_FALSE;
 }
 
 /**
  * Returns true if the rectangle has no area.
  */
-SDL_FORCE_INLINE SDL_bool SDL_RectEmpty(const SDL_Rect *r)
-{
-    return ((!r) || (r->w <= 0) || (r->h <= 0)) ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_RectEmpty(const SDL_Rect *r) {
+  return ((!r) || (r->w <= 0) || (r->h <= 0)) ? SDL_TRUE : SDL_FALSE;
 }
 
 /**
  * Returns true if the two rectangles are equal.
  */
-SDL_FORCE_INLINE SDL_bool SDL_RectEquals(const SDL_Rect *a, const SDL_Rect *b)
-{
-    return (a && b && (a->x == b->x) && (a->y == b->y) &&
-            (a->w == b->w) && (a->h == b->h)) ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_RectEquals(const SDL_Rect *a, const SDL_Rect *b) {
+  return (a && b && (a->x == b->x) && (a->y == b->y) && (a->w == b->w) &&
+          (a->h == b->h))
+             ? SDL_TRUE
+             : SDL_FALSE;
 }
 
 /**
@@ -143,8 +138,8 @@ SDL_FORCE_INLINE SDL_bool SDL_RectEquals(const SDL_Rect *a, const SDL_Rect *b)
  *
  * \sa SDL_IntersectRect
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersection(const SDL_Rect * A,
-                                                     const SDL_Rect * B);
+extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersection(const SDL_Rect *A,
+                                                     const SDL_Rect *B);
 
 /**
  * Calculate the intersection of two rectangles.
@@ -161,9 +156,9 @@ extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersection(const SDL_Rect * A,
  *
  * \sa SDL_HasIntersection
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_IntersectRect(const SDL_Rect * A,
-                                                   const SDL_Rect * B,
-                                                   SDL_Rect * result);
+extern DECLSPEC SDL_bool SDLCALL SDL_IntersectRect(const SDL_Rect *A,
+                                                   const SDL_Rect *B,
+                                                   SDL_Rect *result);
 
 /**
  * Calculate the union of two rectangles.
@@ -175,9 +170,8 @@ extern DECLSPEC SDL_bool SDLCALL SDL_IntersectRect(const SDL_Rect * A,
  *
  * \since This function is available since SDL 2.0.0.
  */
-extern DECLSPEC void SDLCALL SDL_UnionRect(const SDL_Rect * A,
-                                           const SDL_Rect * B,
-                                           SDL_Rect * result);
+extern DECLSPEC void SDLCALL SDL_UnionRect(const SDL_Rect *A, const SDL_Rect *B,
+                                           SDL_Rect *result);
 
 /**
  * Calculate a minimal rectangle enclosing a set of points.
@@ -196,10 +190,10 @@ extern DECLSPEC void SDLCALL SDL_UnionRect(const SDL_Rect * A,
  *
  * \since This function is available since SDL 2.0.0.
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_EnclosePoints(const SDL_Point * points,
+extern DECLSPEC SDL_bool SDLCALL SDL_EnclosePoints(const SDL_Point *points,
                                                    int count,
-                                                   const SDL_Rect * clip,
-                                                   SDL_Rect * result);
+                                                   const SDL_Rect *clip,
+                                                   SDL_Rect *result);
 
 /**
  * Calculate the intersection of a rectangle and line segment.
@@ -219,29 +213,28 @@ extern DECLSPEC SDL_bool SDLCALL SDL_EnclosePoints(const SDL_Point * points,
  *
  * \since This function is available since SDL 2.0.0.
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_IntersectRectAndLine(const SDL_Rect *
-                                                          rect, int *X1,
-                                                          int *Y1, int *X2,
-                                                          int *Y2);
-
+extern DECLSPEC SDL_bool SDLCALL SDL_IntersectRectAndLine(const SDL_Rect *rect,
+                                                          int *X1, int *Y1,
+                                                          int *X2, int *Y2);
 
 /* SDL_FRect versions... */
 
 /**
  * Returns true if point resides inside a rectangle.
  */
-SDL_FORCE_INLINE SDL_bool SDL_PointInFRect(const SDL_FPoint *p, const SDL_FRect *r)
-{
-    return ( (p->x >= r->x) && (p->x < (r->x + r->w)) &&
-             (p->y >= r->y) && (p->y < (r->y + r->h)) ) ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_PointInFRect(const SDL_FPoint *p,
+                                           const SDL_FRect *r) {
+  return ((p->x >= r->x) && (p->x < (r->x + r->w)) && (p->y >= r->y) &&
+          (p->y < (r->y + r->h)))
+             ? SDL_TRUE
+             : SDL_FALSE;
 }
 
 /**
  * Returns true if the rectangle has no area.
  */
-SDL_FORCE_INLINE SDL_bool SDL_FRectEmpty(const SDL_FRect *r)
-{
-    return ((!r) || (r->w <= 0.0f) || (r->h <= 0.0f)) ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_FRectEmpty(const SDL_FRect *r) {
+  return ((!r) || (r->w <= 0.0f) || (r->h <= 0.0f)) ? SDL_TRUE : SDL_FALSE;
 }
 
 /**
@@ -249,14 +242,16 @@ SDL_FORCE_INLINE SDL_bool SDL_FRectEmpty(const SDL_FRect *r)
  *
  * \since This function is available since SDL 2.0.22.
  */
-SDL_FORCE_INLINE SDL_bool SDL_FRectEqualsEpsilon(const SDL_FRect *a, const SDL_FRect *b, const float epsilon)
-{
-    return (a && b && ((a == b) ||
-            ((SDL_fabs(a->x - b->x) <= epsilon) &&
-            (SDL_fabs(a->y - b->y) <= epsilon) &&
-            (SDL_fabs(a->w - b->w) <= epsilon) &&
-            (SDL_fabs(a->h - b->h) <= epsilon))))
-            ? SDL_TRUE : SDL_FALSE;
+SDL_FORCE_INLINE SDL_bool SDL_FRectEqualsEpsilon(const SDL_FRect *a,
+                                                 const SDL_FRect *b,
+                                                 const double epsilon) {
+  return (a && b &&
+          ((a == b) || ((SDL_fabs(a->x - b->x) <= epsilon) &&
+                        (SDL_fabs(a->y - b->y) <= epsilon) &&
+                        (SDL_fabs(a->w - b->w) <= epsilon) &&
+                        (SDL_fabs(a->h - b->h) <= epsilon))))
+             ? SDL_TRUE
+             : SDL_FALSE;
 }
 
 /**
@@ -264,13 +259,13 @@ SDL_FORCE_INLINE SDL_bool SDL_FRectEqualsEpsilon(const SDL_FRect *a, const SDL_F
  *
  * \since This function is available since SDL 2.0.22.
  */
-SDL_FORCE_INLINE SDL_bool SDL_FRectEquals(const SDL_FRect *a, const SDL_FRect *b)
-{
-    return SDL_FRectEqualsEpsilon(a, b, SDL_FLT_EPSILON);
+SDL_FORCE_INLINE SDL_bool SDL_FRectEquals(const SDL_FRect *a,
+                                          const SDL_FRect *b) {
+  return SDL_FRectEqualsEpsilon(a, b, SDL_FLT_EPSILON);
 }
 
 /**
- * Determine whether two rectangles intersect with float precision.
+ * Determine whether two rectangles intersect with double precision.
  *
  * If either pointer is NULL the function will return SDL_FALSE.
  *
@@ -282,11 +277,11 @@ SDL_FORCE_INLINE SDL_bool SDL_FRectEquals(const SDL_FRect *a, const SDL_FRect *b
  *
  * \sa SDL_IntersectRect
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersectionF(const SDL_FRect * A,
-                                                      const SDL_FRect * B);
+extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersectionF(const SDL_FRect *A,
+                                                      const SDL_FRect *B);
 
 /**
- * Calculate the intersection of two rectangles with float precision.
+ * Calculate the intersection of two rectangles with double precision.
  *
  * If `result` is NULL then this function will return SDL_FALSE.
  *
@@ -300,12 +295,12 @@ extern DECLSPEC SDL_bool SDLCALL SDL_HasIntersectionF(const SDL_FRect * A,
  *
  * \sa SDL_HasIntersectionF
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_IntersectFRect(const SDL_FRect * A,
-                                                    const SDL_FRect * B,
-                                                    SDL_FRect * result);
+extern DECLSPEC SDL_bool SDLCALL SDL_IntersectFRect(const SDL_FRect *A,
+                                                    const SDL_FRect *B,
+                                                    SDL_FRect *result);
 
 /**
- * Calculate the union of two rectangles with float precision.
+ * Calculate the union of two rectangles with double precision.
  *
  * \param A an SDL_FRect structure representing the first rectangle
  * \param B an SDL_FRect structure representing the second rectangle
@@ -314,12 +309,12 @@ extern DECLSPEC SDL_bool SDLCALL SDL_IntersectFRect(const SDL_FRect * A,
  *
  * \since This function is available since SDL 2.0.22.
  */
-extern DECLSPEC void SDLCALL SDL_UnionFRect(const SDL_FRect * A,
-                                            const SDL_FRect * B,
-                                            SDL_FRect * result);
+extern DECLSPEC void SDLCALL SDL_UnionFRect(const SDL_FRect *A,
+                                            const SDL_FRect *B,
+                                            SDL_FRect *result);
 
 /**
- * Calculate a minimal rectangle enclosing a set of points with float
+ * Calculate a minimal rectangle enclosing a set of points with double
  * precision.
  *
  * If `clip` is not NULL then only points inside of the clipping rectangle are
@@ -336,13 +331,13 @@ extern DECLSPEC void SDLCALL SDL_UnionFRect(const SDL_FRect * A,
  *
  * \since This function is available since SDL 2.0.22.
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_EncloseFPoints(const SDL_FPoint * points,
+extern DECLSPEC SDL_bool SDLCALL SDL_EncloseFPoints(const SDL_FPoint *points,
                                                     int count,
-                                                    const SDL_FRect * clip,
-                                                    SDL_FRect * result);
+                                                    const SDL_FRect *clip,
+                                                    SDL_FRect *result);
 
 /**
- * Calculate the intersection of a rectangle and line segment with float
+ * Calculate the intersection of a rectangle and line segment with double
  * precision.
  *
  * This function is used to clip a line segment to a rectangle. A line segment
@@ -360,10 +355,8 @@ extern DECLSPEC SDL_bool SDLCALL SDL_EncloseFPoints(const SDL_FPoint * points,
  *
  * \since This function is available since SDL 2.0.22.
  */
-extern DECLSPEC SDL_bool SDLCALL SDL_IntersectFRectAndLine(const SDL_FRect *
-                                                           rect, float *X1,
-                                                           float *Y1, float *X2,
-                                                           float *Y2);
+extern DECLSPEC SDL_bool SDLCALL SDL_IntersectFRectAndLine(
+    const SDL_FRect *rect, double *X1, double *Y1, double *X2, double *Y2);
 
 /* Ends C function definitions when using C++ */
 #ifdef __cplusplus
