+++ src/SDL2/x86_64-w64-mingw32/include/SDL2/SDL_sensor.h
@@ -29,8 +29,8 @@
 #ifndef SDL_sensor_h_
 #define SDL_sensor_h_
 
-#include "SDL_stdinc.h"
 #include "SDL_error.h"
+#include "SDL_stdinc.h"
 
 #include "begin_code.h"
 /* Set up for C function definitions, even when using C++ */
@@ -55,7 +55,8 @@ typedef struct _SDL_Sensor SDL_Sensor;
  * This is a unique ID for a sensor for the time it is connected to the system,
  * and is never reused for the lifetime of the application.
  *
- * The ID value starts at 0 and increments from there. The value -1 is an invalid ID.
+ * The ID value starts at 0 and increments from there. The value -1 is an
+ * invalid ID.
  */
 typedef Sint32 SDL_SensorID;
 
@@ -66,12 +67,11 @@ typedef Sint32 SDL_SensorID;
  * Hare are the additional Android sensors:
  * https://developer.android.com/reference/android/hardware/SensorEvent.html#values
  */
-typedef enum
-{
-    SDL_SENSOR_INVALID = -1,    /**< Returned for an invalid sensor */
-    SDL_SENSOR_UNKNOWN,         /**< Unknown sensor type */
-    SDL_SENSOR_ACCEL,           /**< Accelerometer */
-    SDL_SENSOR_GYRO             /**< Gyroscope */
+typedef enum {
+  SDL_SENSOR_INVALID = -1, /**< Returned for an invalid sensor */
+  SDL_SENSOR_UNKNOWN,      /**< Unknown sensor type */
+  SDL_SENSOR_ACCEL,        /**< Accelerometer */
+  SDL_SENSOR_GYRO          /**< Gyroscope */
 } SDL_SensorType;
 
 /**
@@ -91,12 +91,12 @@ typedef enum
  * -X ... +X : left ... right
  * -Y ... +Y : bottom ... top
  * -Z ... +Z : farther ... closer
- * 
+ *
  * The axis data is not changed when the phone is rotated.
  *
  * \sa SDL_GetDisplayOrientation()
  */
-#define SDL_STANDARD_GRAVITY    9.80665f
+#define SDL_STANDARD_GRAVITY 9.80665f
 
 /**
  * Gyroscope sensor
@@ -116,7 +116,7 @@ typedef enum
  * -X ... +X : left ... right
  * -Y ... +Y : bottom ... top
  * -Z ... +Z : farther ... closer
- * 
+ *
  * The axis data is not changed when the phone or controller is rotated.
  *
  * \sa SDL_GetDisplayOrientation()
@@ -167,7 +167,8 @@ extern DECLSPEC const char *SDLCALL SDL_SensorGetDeviceName(int device_index);
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC SDL_SensorType SDLCALL SDL_SensorGetDeviceType(int device_index);
+extern DECLSPEC SDL_SensorType SDLCALL
+SDL_SensorGetDeviceType(int device_index);
 
 /**
  * Get the platform dependent type of a sensor.
@@ -178,7 +179,8 @@ extern DECLSPEC SDL_SensorType SDLCALL SDL_SensorGetDeviceType(int device_index)
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC int SDLCALL SDL_SensorGetDeviceNonPortableType(int device_index);
+extern DECLSPEC int SDLCALL
+SDL_SensorGetDeviceNonPortableType(int device_index);
 
 /**
  * Get the instance ID of a sensor.
@@ -188,7 +190,8 @@ extern DECLSPEC int SDLCALL SDL_SensorGetDeviceNonPortableType(int device_index)
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC SDL_SensorID SDLCALL SDL_SensorGetDeviceInstanceID(int device_index);
+extern DECLSPEC SDL_SensorID SDLCALL
+SDL_SensorGetDeviceInstanceID(int device_index);
 
 /**
  * Open a sensor for use.
@@ -208,7 +211,8 @@ extern DECLSPEC SDL_Sensor *SDLCALL SDL_SensorOpen(int device_index);
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC SDL_Sensor *SDLCALL SDL_SensorFromInstanceID(SDL_SensorID instance_id);
+extern DECLSPEC SDL_Sensor *SDLCALL
+SDL_SensorFromInstanceID(SDL_SensorID instance_id);
 
 /**
  * Get the implementation dependent name of a sensor
@@ -249,7 +253,8 @@ extern DECLSPEC int SDLCALL SDL_SensorGetNonPortableType(SDL_Sensor *sensor);
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC SDL_SensorID SDLCALL SDL_SensorGetInstanceID(SDL_Sensor *sensor);
+extern DECLSPEC SDL_SensorID SDLCALL
+SDL_SensorGetInstanceID(SDL_Sensor *sensor);
 
 /**
  * Get the current state of an opened sensor.
@@ -263,7 +268,8 @@ extern DECLSPEC SDL_SensorID SDLCALL SDL_SensorGetInstanceID(SDL_Sensor *sensor)
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC int SDLCALL SDL_SensorGetData(SDL_Sensor * sensor, float *data, int num_values);
+extern DECLSPEC int SDLCALL SDL_SensorGetData(SDL_Sensor *sensor, double *data,
+                                              int num_values);
 
 /**
  * Close a sensor previously opened with SDL_SensorOpen().
@@ -272,7 +278,7 @@ extern DECLSPEC int SDLCALL SDL_SensorGetData(SDL_Sensor * sensor, float *data, 
  *
  * \since This function is available since SDL 2.0.9.
  */
-extern DECLSPEC void SDLCALL SDL_SensorClose(SDL_Sensor * sensor);
+extern DECLSPEC void SDLCALL SDL_SensorClose(SDL_Sensor *sensor);
 
 /**
  * Update the current state of the open sensors.
@@ -286,7 +292,6 @@ extern DECLSPEC void SDLCALL SDL_SensorClose(SDL_Sensor * sensor);
  * \since This function is available since SDL 2.0.9.
  */
 extern DECLSPEC void SDLCALL SDL_SensorUpdate(void);
-
 
 /* Ends C function definitions when using C++ */
 #ifdef __cplusplus
