+++ src/SDL2/x86_64-w64-mingw32/include/SDL2/SDL_test_common.h
@@ -35,88 +35,87 @@
 #include "SDL.h"
 
 #if defined(__PSP__)
-#define DEFAULT_WINDOW_WIDTH  480
+#define DEFAULT_WINDOW_WIDTH 480
 #define DEFAULT_WINDOW_HEIGHT 272
 #elif defined(__VITA__)
-#define DEFAULT_WINDOW_WIDTH  960
+#define DEFAULT_WINDOW_WIDTH 960
 #define DEFAULT_WINDOW_HEIGHT 544
 #else
-#define DEFAULT_WINDOW_WIDTH  640
+#define DEFAULT_WINDOW_WIDTH 640
 #define DEFAULT_WINDOW_HEIGHT 480
 #endif
 
-#define VERBOSE_VIDEO   0x00000001
-#define VERBOSE_MODES   0x00000002
-#define VERBOSE_RENDER  0x00000004
-#define VERBOSE_EVENT   0x00000008
-#define VERBOSE_AUDIO   0x00000010
+#define VERBOSE_VIDEO 0x00000001
+#define VERBOSE_MODES 0x00000002
+#define VERBOSE_RENDER 0x00000004
+#define VERBOSE_EVENT 0x00000008
+#define VERBOSE_AUDIO 0x00000010
 
-typedef struct
-{
-    /* SDL init flags */
-    char **argv;
-    Uint32 flags;
-    Uint32 verbose;
+typedef struct {
+  /* SDL init flags */
+  char **argv;
+  Uint32 flags;
+  Uint32 verbose;
 
-    /* Video info */
-    const char *videodriver;
-    int display;
-    const char *window_title;
-    const char *window_icon;
-    Uint32 window_flags;
-    SDL_bool flash_on_focus_loss;
-    int window_x;
-    int window_y;
-    int window_w;
-    int window_h;
-    int window_minW;
-    int window_minH;
-    int window_maxW;
-    int window_maxH;
-    int logical_w;
-    int logical_h;
-    float scale;
-    int depth;
-    int refresh_rate;
-    int num_windows;
-    SDL_Window **windows;
+  /* Video info */
+  const char *videodriver;
+  int display;
+  const char *window_title;
+  const char *window_icon;
+  Uint32 window_flags;
+  SDL_bool flash_on_focus_loss;
+  int window_x;
+  int window_y;
+  int window_w;
+  int window_h;
+  int window_minW;
+  int window_minH;
+  int window_maxW;
+  int window_maxH;
+  int logical_w;
+  int logical_h;
+  double scale;
+  int depth;
+  int refresh_rate;
+  int num_windows;
+  SDL_Window **windows;
 
-    /* Renderer info */
-    const char *renderdriver;
-    Uint32 render_flags;
-    SDL_bool skip_renderer;
-    SDL_Renderer **renderers;
-    SDL_Texture **targets;
+  /* Renderer info */
+  const char *renderdriver;
+  Uint32 render_flags;
+  SDL_bool skip_renderer;
+  SDL_Renderer **renderers;
+  SDL_Texture **targets;
 
-    /* Audio info */
-    const char *audiodriver;
-    SDL_AudioSpec audiospec;
+  /* Audio info */
+  const char *audiodriver;
+  SDL_AudioSpec audiospec;
 
-    /* GL settings */
-    int gl_red_size;
-    int gl_green_size;
-    int gl_blue_size;
-    int gl_alpha_size;
-    int gl_buffer_size;
-    int gl_depth_size;
-    int gl_stencil_size;
-    int gl_double_buffer;
-    int gl_accum_red_size;
-    int gl_accum_green_size;
-    int gl_accum_blue_size;
-    int gl_accum_alpha_size;
-    int gl_stereo;
-    int gl_multisamplebuffers;
-    int gl_multisamplesamples;
-    int gl_retained_backing;
-    int gl_accelerated;
-    int gl_major_version;
-    int gl_minor_version;
-    int gl_debug;
-    int gl_profile_mask;
+  /* GL settings */
+  int gl_red_size;
+  int gl_green_size;
+  int gl_blue_size;
+  int gl_alpha_size;
+  int gl_buffer_size;
+  int gl_depth_size;
+  int gl_stencil_size;
+  int gl_double_buffer;
+  int gl_accum_red_size;
+  int gl_accum_green_size;
+  int gl_accum_blue_size;
+  int gl_accum_alpha_size;
+  int gl_stereo;
+  int gl_multisamplebuffers;
+  int gl_multisamplesamples;
+  int gl_retained_backing;
+  int gl_accelerated;
+  int gl_major_version;
+  int gl_minor_version;
+  int gl_debug;
+  int gl_profile_mask;
 
-    /* Additional fields added in 2.0.18 */
-    SDL_Rect confine;
+  /* Additional fields added in 2.0.18 */
+  SDL_Rect confine;
 
 } SDLTest_CommonState;
 
@@ -132,7 +131,8 @@ extern "C" {
  * \brief Parse command line parameters and create common state.
  *
  * \param argv Array of command line parameters
- * \param flags Flags indicating which subsystem to initialize (i.e. SDL_INIT_VIDEO | SDL_INIT_AUDIO)
+ * \param flags Flags indicating which subsystem to initialize (i.e.
+ * SDL_INIT_VIDEO | SDL_INIT_AUDIO)
  *
  * \returns a newly allocated common state object.
  */
@@ -144,10 +144,10 @@ SDLTest_CommonState *SDLTest_CommonCreateState(char **argv, Uint32 flags);
  * \param state The common state describing the test window to create.
  * \param index The index of the argument to process in argv[].
  *
- * \returns the number of arguments processed (i.e. 1 for --fullscreen, 2 for --video [videodriver], or -1 on error.
+ * \returns the number of arguments processed (i.e. 1 for --fullscreen, 2 for
+ * --video [videodriver], or -1 on error.
  */
-int SDLTest_CommonArg(SDLTest_CommonState * state, int index);
-
+int SDLTest_CommonArg(SDLTest_CommonState *state, int index);
 
 /**
  * \brief Logs command line usage info.
@@ -159,9 +159,11 @@ int SDLTest_CommonArg(SDLTest_CommonState * state, int index);
  *
  * \param state The common state describing the test window for the app.
  * \param argv0 argv[0], as passed to main/SDL_main.
- * \param options an array of strings for application specific options. The last element of the array should be NULL.
+ * \param options an array of strings for application specific options. The last
+ * element of the array should be NULL.
  */
-void SDLTest_CommonLogUsage(SDLTest_CommonState * state, const char *argv0, const char **options);
+void SDLTest_CommonLogUsage(SDLTest_CommonState *state, const char *argv0,
+                            const char **options);
 
 /**
  * \brief Returns common usage information
@@ -174,7 +176,7 @@ void SDLTest_CommonLogUsage(SDLTest_CommonState * state, const char *argv0, cons
  * \param state The common state describing the test window to create.
  * \returns a string with usage information
  */
-const char *SDLTest_CommonUsage(SDLTest_CommonState * state);
+const char *SDLTest_CommonUsage(SDLTest_CommonState *state);
 
 /**
  * \brief Open test window.
@@ -183,7 +185,7 @@ const char *SDLTest_CommonUsage(SDLTest_CommonState * state);
  *
  * \returns SDL_TRUE if initialization succeeded, false otherwise
  */
-SDL_bool SDLTest_CommonInit(SDLTest_CommonState * state);
+SDL_bool SDLTest_CommonInit(SDLTest_CommonState *state);
 
 /**
  * \brief Easy argument handling when test app doesn't need any custom args.
@@ -194,7 +196,8 @@ SDL_bool SDLTest_CommonInit(SDLTest_CommonState * state);
  *
  * \returns SDL_FALSE if app should quit, true otherwise.
  */
-SDL_bool SDLTest_CommonDefaultArgs(SDLTest_CommonState * state, const int argc, char **argv);
+SDL_bool SDLTest_CommonDefaultArgs(SDLTest_CommonState *state, const int argc,
+                                   char **argv);
 
 /**
  * \brief Common event handler for test windows.
@@ -204,7 +207,8 @@ SDL_bool SDLTest_CommonDefaultArgs(SDLTest_CommonState * state, const int argc, 
  * \param done Flag indicating we are done.
  *
  */
-void SDLTest_CommonEvent(SDLTest_CommonState * state, SDL_Event * event, int *done);
+void SDLTest_CommonEvent(SDLTest_CommonState *state, SDL_Event *event,
+                         int *done);
 
 /**
  * \brief Close test window.
@@ -212,17 +216,19 @@ void SDLTest_CommonEvent(SDLTest_CommonState * state, SDL_Event * event, int *do
  * \param state The common state used to create test window.
  *
  */
-void SDLTest_CommonQuit(SDLTest_CommonState * state);
+void SDLTest_CommonQuit(SDLTest_CommonState *state);
 
 /**
- * \brief Draws various window information (position, size, etc.) to the renderer.
+ * \brief Draws various window information (position, size, etc.) to the
+ * renderer.
  *
  * \param renderer The renderer to draw to.
  * \param window The window whose information should be displayed.
  * \param usedHeight Returns the height used, so the caller can draw more below.
  *
  */
-void SDLTest_CommonDrawWindowInfo(SDL_Renderer * renderer, SDL_Window * window, int * usedHeight);
+void SDLTest_CommonDrawWindowInfo(SDL_Renderer *renderer, SDL_Window *window,
+                                  int *usedHeight);
 
 /* Ends C function definitions when using C++ */
 #ifdef __cplusplus
